[cli][mcp] expose cli extensions as mcp tools
CLI extensions are already supported by the CLI, where any expo package can expose tools that will be available in the CLI "more tools" menu.

These commands are suitable as MCP tools as well, since they follow a fixed schema for input/output.

This PR implements support for exposing a command in a cli extension as an MCP tool.